+++ examples/fnft_nsep_example.c
@@ -81,9 +81,9 @@ int main()
 
     // See the header file fnft_nsep.h for other options
 
-    /** Step 3: Call fnft_nsev and check for errors **/
+    /** Step 3: Call fnft_nsep and check for errors **/
 
-    int ret_code = fnft_nsep(D, q, T, phase_shift, &K, main_spec, &M, aux_spec, NULL, kappa,
+    FNFT_INT ret_code = fnft_nsep(D, q, T, phase_shift, &K, main_spec, &M, aux_spec, NULL, kappa,
         &opts);
     if (ret_code != FNFT_SUCCESS) {
         printf("An error occured!\n");