+++ matlab/mex_fnft_nsep.c
@@ -68,7 +68,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[])
 
     /* Check values of first three inputs */
     if ( D<2 || D%2 != 0 )
-        mexErrMsgTxt("Length of the first input q should be odd.");
+        mexErrMsgTxt("Length of the first input q should be even.");
     if ( T[0] >= T[1] )
         mexErrMsgTxt("T(1) >= T(2).");
     if ( kappa != +1 && kappa != -1 )