+++ src/private/fnft__akns_scatter_matrix.c
@@ -197,7 +197,7 @@ INT akns_scatter_matrix(const UINT D, COMPLEX const * const q,
                 U[3][2] = rn*sh;
                 U[3][3] = ch+u1;
                 
-                misc_square_matrix_mult(4, &U[0][0], &T[0][0]);
+                misc_mat_mult_4x4(&U[0][0], &T[0][0]);
             }
             
             result[i*8] = T[0][0];
@@ -230,7 +230,7 @@ INT akns_scatter_matrix(const UINT D, COMPLEX const * const q,
                 U[0][1] = qn*sh;
                 U[1][0] = rn*sh;
                 U[1][1] = ch + u1;
-                misc_square_matrix_mult(2, &U[0][0], &T[0][0]);
+                misc_mat_mult_2x2(&U[0][0], &T[0][0]);
             }
             
             result[i*4] = T[0][0];
@@ -348,7 +348,7 @@ INT akns_scatter_matrix(const UINT D, COMPLEX const * const q,
                             U[2][3] = U[0][1];
                             U[3][2] = U[1][0];
                             U[3][3] = U[1][1];
-                            misc_square_matrix_mult(4, &U[0][0], &T[0][0]);
+                            misc_mat_mult_4x4(&U[0][0], &T[0][0]);
                         }
                         break;
                         // Fourth-order exponential method which requires
@@ -396,8 +396,8 @@ INT akns_scatter_matrix(const UINT D, COMPLEX const * const q,
                             UD[1][0] = w_d*r[n];
                             UD[1][1] = c_d+I*s_d;
                             
-                            misc_square_matrix_mult(2, &UN[0][0], &TM[0][0]);
-                            misc_square_matrix_mult(2, &UD[0][0], &TMD[0][0]);
+                            misc_mat_mult_2x2(&UN[0][0], &TM[0][0]);
+                            misc_mat_mult_2x2(&UD[0][0], &TMD[0][0]);
                             
                             a1 = tmp2[n];
                             a2 = tmp2[n+1];
@@ -417,8 +417,8 @@ INT akns_scatter_matrix(const UINT D, COMPLEX const * const q,
                             UD[1][0] = UN[1][0];
                             UD[1][1] = UN[1][1];
                             
-                            misc_square_matrix_mult(2, &UN[0][0], &TM[0][0]);
-                            misc_square_matrix_mult(2, &UD[0][0], &TMD[0][0]);
+                            misc_mat_mult_2x2(&UN[0][0], &TM[0][0]);
+                            misc_mat_mult_2x2(&UD[0][0], &TMD[0][0]);
                             
                             U[0][0] = TM[0][0];
                             U[0][1] = TM[0][1];
@@ -433,7 +433,7 @@ INT akns_scatter_matrix(const UINT D, COMPLEX const * const q,
                             U[3][2] = U[1][0];
                             U[3][3] = U[1][1];
                             
-                            misc_square_matrix_mult(4, &U[0][0], &T[0][0]);
+                            misc_mat_mult_4x4(&U[0][0], &T[0][0]);
                         }
                         break;
                         
@@ -476,7 +476,7 @@ INT akns_scatter_matrix(const UINT D, COMPLEX const * const q,
                             U[0][1] = s*(a1-I*a2);
                             U[1][0] = s*(a1+I*a2);
                             U[1][1] = (c-s*a3);
-                            misc_square_matrix_mult(2, &U[0][0], &T[0][0]);
+                            misc_mat_mult_2x2(&U[0][0], &T[0][0]);
                         }
                         break;
                         // Fourth-order exponential method which requires
@@ -510,7 +510,7 @@ INT akns_scatter_matrix(const UINT D, COMPLEX const * const q,
                             U[0][1] = s*(a1-I*a2);
                             U[1][0] = s*(a1+I*a2);
                             U[1][1] = (c-s*a3);
-                            misc_square_matrix_mult(2, &U[0][0], &T[0][0]);
+                            misc_mat_mult_2x2(&U[0][0], &T[0][0]);
                         }
                         break;